add frontend

+++ ToDo/tasks/views.py
@@ -1,4 +1,5 @@
 from rest_framework.decorators import api_view
+from django.shortcuts import render, redirect
 from rest_framework.response import Response
 from .serializers import TaskSerializer
 from rest_framework import status
@@ -13,7 +14,7 @@ def get_all_tasks(request):
     tasks = Task.objects.all().filter(owner=user).order_by('created_date')
     serializers = TaskSerializer(tasks, many=True)
 
-    return Response(serializers.data)
+    return render(request, 'task_list.html', {'tasks': serializers.data})
 
 
 @api_view(['GET'])
@@ -26,46 +27,50 @@ def get_task(request, pk):
             raise Exception
 
         serializer = TaskSerializer(task, many=False)
-        return Response(serializer.data)
+        return render(request, 'task_detail.html', {'task': serializer.data})
     except:
         message = {'detail': 'No task found with this id'}
-        return Response(message, status=status.HTTP_404_NOT_FOUND)
+        return render(request, '404.html', message)
 
 
-@api_view(['DELETE'])
 def remove_task(request, pk):
     try:
         user = request.user
-        task = task.objects.get(pk=pk)
+        task = Task.objects.get(pk=pk)
+        serializer = TaskSerializer(task, many=False)
 
         if task.owner != user:
             raise Exception
 
-        task.delete()
+        if request.method == 'POST':
+            task.delete()
+            return redirect(get_all_tasks)
 
-        message = {'detail': 'Task deleted'}
-        return Response(message, status=status.HTTP_204_NO_CONTENT)
+        return render(request, 'remove_task.html', {'task': serializer.data})
+
     except:
         message = {'detail': 'No task found with this id'}
-        return Response(message, status=status.HTTP_404_NOT_FOUND)
+        return render(request, '404.html', message)
 
 
-@api_view(['POST'])
 def add_task(request):
-    try:
-        owner = request.user
-        title = request.data.get('title')
-        description = request.data.get('description')
+    if request.method == 'POST':
+        try:
+            owner = request.user
+            title = request.POST.get('title')
+            description = request.POST.get('description')
 
-        task = Task.objects.create(
-            owner = owner,
-            title = title,
-            description = description,
-        )
+            task = Task.objects.create(
+                owner = owner,
+                title = title,
+                description = description,
+            )
 
-        task.save()
-        serializer = TaskSerializer(task, many=False)
-        return Response(serializer.data)
-    except:
-        message = {'detail': 'Can not add task this time'}
-        return Response(message, status=status.HTTP_404_NOT_FOUND)
+            task.save()
+            return redirect(get_all_tasks)
+
+        except:
+            message = {'detail': 'Can not add task this time'}
+            return render(request, '404.html', message)
+
+    return render(request, 'add_task.html')
